move Font out of core
It uses ThreeMeshUI
