+++ src/Font.js
@@ -1,5 +1,5 @@
 import fetch from 'cross-fetch'
-import ThreeMeshUI from '../extensions/ThreeMeshUI.js'
+import ThreeMeshUI from './extensions/ThreeMeshUI.js'
 
 const FETCH_OPTIONS = { method: 'get' }
 const NOOP = () => {}
