+++ src/types/core.js
@@ -2,7 +2,6 @@
  * @typedef { import("../core/Client").Client<*,*,*,*> } Client
  * @typedef { import("../core/ClientData").ClientData } ClientData
  * @typedef { import("../core/ClientAudio").ClientAudio<*> } ClientAudio
- * @typedef { import("../core/Font").Font } Font
  * @typedef { import("../core/Pose").Pose } Pose
  * @typedef { import("../core/Vector3").Vector3 } Vector3
  */
