+++ src/types/index.js
@@ -1,6 +1,7 @@
 /**
  * @typedef {number|string} id - The ID of a resource
  *
+ * @typedef { import("../Font").Font } Font
  * @typedef { import("../ThreeClient").ThreeClient } ThreeClient
  * @typedef { import("../ThreeClientAudio").ThreeClientAudio } ThreeClientAudio
  */
